구독,expert탭수정

+++ src/main/java/com/app/globalgates/dto/PaymentSubscribeDTO.java
@@ -12,7 +12,7 @@ public class PaymentSubscribeDTO {
     private Long id;
     private Long subscriptionId;
     private Long memberId;
-    private Long amount;
+    private int amount;
     private PaymentStatus paymentStatus;
     private String paymentMethod;
     private String receiptId;
